+++ plugins/csp-fly-to-locations/src/Plugin.cpp
@@ -33,11 +33,8 @@ void Plugin::init() {
 
   logger().info("Loading plugin...");
 
-  mGuiManager->addHtmlToGui(
-      "fly-to-locations", "../share/resources/gui/fly-to-locations-templates.html");
-
-  mGuiManager->addScriptToGuiFromJS("../share/resources/gui/js/csp-fly-to-locations.js");
-  mGuiManager->addCssToGui("css/csp-fly-to-locations.css");
+  mGuiManager->executeJavascriptFile("../share/resources/gui/js/csp-fly-to-locations.js");
+  mGuiManager->addCSS("css/csp-fly-to-locations.css");
 
   mGuiManager->addPluginTabToSideBarFromHTML(
       "Bookmarks", "place", "../share/resources/gui/fly-to-locations-tab.html");
@@ -93,9 +90,7 @@ void Plugin::deInit() {
   mGuiManager->onBookmarkAdded().disconnect(mOnBookmarkAddedConnection);
   mGuiManager->onBookmarkRemoved().disconnect(mOnBookmarkRemovedConnection);
 
-  mGuiManager->getGui()->callJavascript("CosmoScout.gui.unregisterHtml", "fly-to-locations");
-  mGuiManager->getGui()->callJavascript(
-      "CosmoScout.gui.unregisterCss", "css/csp-fly-to-locations.css");
+  mGuiManager->removeCSS("css/csp-fly-to-locations.css");
 
   mSolarSystem->pActiveObject.disconnect(mActiveBodyConnection);