+++ resources/gui/js/apis/notification.js
@@ -67,9 +67,8 @@ class NotificationApi extends IApi {
 
     if (flyTo) {
       notification.classList.add('clickable');
-      notification.addEventListener('click', () => {
-        CosmoScout.callbacks.navigation.flyTo(flyTo);
-      });
+      notification.addEventListener(
+          'click', () => { CosmoScout.callbacks.navigation.flyTo(flyTo); });
     }
 
     notification.timer = setTimeout(() => {
@@ -77,9 +76,7 @@ class NotificationApi extends IApi {
       this._container.removeChild(notification);
     }, 8000);
 
-    setTimeout(() => {
-      notification.classList.add('show');
-    }, 60);
+    setTimeout(() => { notification.classList.add('show'); }, 60);
   }
 
   /**
@@ -92,7 +89,7 @@ class NotificationApi extends IApi {
    * @private
    */
   _makeNotification(title, content, icon = '') {
-    const notification = CosmoScout.gui.loadTemplateContent('notification');
+    const notification = CosmoScout.gui.loadTemplateContent('notification-template');
 
     if (notification === false) {
       throw new Error(