+++ resources/gui/js/apis/timeline.js
@@ -217,12 +217,10 @@ class TimelineApi extends IApi {
    * @param callback {string} Name of callback on CosmoScout.callbacks
    */
   addButton(name, icon, callback) {
-    const button     = CosmoScout.gui.loadTemplateContent('button');
+    const button     = CosmoScout.gui.loadTemplateContent('button-template');
     button.innerHTML = button.innerHTML.replace('%ICON%', icon).trim();
     button.setAttribute('title', name);
-    button.onclick = () => {
-      CosmoScout.callbacks.find(callback)();
-    };
+    button.onclick = () => { CosmoScout.callbacks.find(callback)(); };
 
     this._buttonContainer.appendChild(button);